Move http client to it's own file; Add ability to clone HTTP client for per-request configuration changes.
