+++ lib/web/apiserver.go
@@ -3117,6 +3117,10 @@ func (h *Handler) createSSHCert(w http.ResponseWriter, r *http.Request, p httpro
 		return nil, trace.Wrap(err)
 	}
 
+	var authenticationClient interface {
+		AuthenticateSSHUser(ctx context.Context, req auth.AuthenticateSSHRequest) (*auth.SSHLoginResponse, error)
+	} = authClient
+
 	authReq := auth.AuthenticateUserRequest{
 		Username:       req.User,
 		PublicKey:      req.PubKey,
@@ -3136,11 +3140,26 @@ func (h *Handler) createSSHCert(w http.ResponseWriter, r *http.Request, p httpro
 	case constants.SecondFactorWebauthn:
 		// WebAuthn only supports this endpoint for headless login.
 		authReq.HeadlessAuthenticationID = req.HeadlessAuthenticationID
+
+		clt, ok := authClient.(*auth.Client)
+		if !ok {
+			return nil, trace.Errorf("expected client type *auth.Client but got %T", authClient)
+		}
+
+		clientParams := []roundtrip.ClientParam{
+			auth.ClientParamTimeout(defaults.CallbackTimeout),
+			auth.ClientParamResponseHeaderTimeout(defaults.CallbackTimeout),
+		}
+
+		authenticationClient, err = clt.HTTPClient.Clone(clientParams...)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
 	default:
 		return nil, trace.AccessDenied("unsupported second factor type: %q", cap.GetSecondFactor())
 	}
 
-	loginResp, err := authClient.AuthenticateSSHUser(r.Context(), auth.AuthenticateSSHRequest{
+	loginResp, err := authenticationClient.AuthenticateSSHUser(r.Context(), auth.AuthenticateSSHRequest{
 		AuthenticateUserRequest: authReq,
 		CompatibilityMode:       req.Compatibility,
 		TTL:                     req.TTL,
